Response to CR
